bytemymood/requirements.txt
google-cloud-aiplatform[adk,agent-engines]>=1.93.0,<2.0.0
pydantic>=2.10.6,<3.0.0
python-dotenv>=1.0.1,<2.0.0
google-genai>=1.16.1,<2.0.0
google-adk>=1.0.0,<2.0.0
Pillow>=10.0.0,<11.0.0
pytest>=8.3.5,<9.0.0
google-adk[eval]>=1.0.0,<2.0.0
pytest-asyncio>=0.26.0,<0.27.0
absl-py>=2.2.1,<3.0.0
cloudpickle>=3.1.1,<4.0.0
flake8-pyproject>=1.2.3,<2.0.0 